Make use of partial class to reduce file size, add hardcoded monster look types
